App data controller refactor (#24)
* unused service removed

* extracted lastUpdate endpoint to another Coctroller
